Sector time now a double.

+++ iRacingSimulator/Sector.cs
@@ -11,7 +11,7 @@ namespace iRacingSimulator
         public int Number { get; set; }
         public float StartPercentage { get; set; }
 
-        public float EnterSessionTime { get; set; }
+        public double EnterSessionTime { get; set; }
         public Laptime SectorTime { get; set; }
 
         public Sector Copy()
